Solve 1010.c Problem

+++ 1010.c
@@ -1,17 +1,16 @@
 #include <stdio.h>
 
-
 int main() {
     //Initialize variable
     int codeOfProduct, unitOfProduct;
-    float amountOfPerProduct, totalAmountOfProduct = 0;
+    double amountOfPerProduct, totalAmountOfProduct = 0;
     for(int i = 1; i <= 2; i++){
         //take input
-        scanf("%d %d %f",&codeOfProduct, &unitOfProduct, &amountOfPerProduct);
+        scanf("%d %d %lf",&codeOfProduct, &unitOfProduct, &amountOfPerProduct);
         totalAmountOfProduct += unitOfProduct * amountOfPerProduct;
     }
     //output the total amount
-    printf("VALOR A PAGAR: R$ %.2f\n",totalAmountOfProduct);
+    printf("VALOR A PAGAR: R$ %.2lf\n",totalAmountOfProduct);
 
     return 0;
 }
